feat: add github workflows

+++ src/utils/copyTemplate.ts
@@ -3,15 +3,15 @@ import { fileURLToPath } from "node:url"
 import fs from "fs-extra";
 import ora from "ora";
 import chalk from "chalk";
+import { githubWorkflowPath } from "./constant";
 
 export const copyTemplate = async (result: string[]) => {
     const _filename = fileURLToPath(import.meta.url);
     const _dirname = dirname(_filename);
-    const files = ['.gitignore'];
+    const files = ['.github'];
     const targetDirectory = path.resolve(_dirname, '../templates');
     const dest = path.resolve(process.cwd());
-    const spinner = ora('Copying template...').start();
-    const startTime: number = new Date().getTime();
+    const spinner = ora();
     const templates = files.filter(file => !result.includes(file));
 
     async function copy() {
@@ -19,34 +19,26 @@ export const copyTemplate = async (result: string[]) => {
             if (!dest) {
                 return;
             };
+
+            if(fs.existsSync(githubWorkflowPath)) {
+                return;
+            }
             await Promise.all(templates.map(async (file) => {
                 const srcPath = path.join(targetDirectory, file);
                 const destPath = path.join(dest, file);
                 const srcStats = await fs.stat(srcPath);
                 if (srcStats.isFile()) {
                     await fs.copy(srcPath, destPath);
-                };
+                } else {
+                    await fs.copy(srcPath, destPath);
+                }
             }));
+            spinner.succeed('Github workflows template added successfully!')
         } catch (error) {
             if(error instanceof Error) {
                 return spinner.fail(chalk.red(error.message));
             }
-        } finally {
-            spinner.stop();
         }
     }
     await copy();
-
-    if(!templates.length) {
-        return;
-    } else {
-        spinner.succeed(chalk.green('Template(s) added successfully!'))
-    };
-
-    const endTime: number = new Date().getTime()
-    const usageTime: number = (endTime - startTime) / 1000;
-
-    console.log();
-
-    console.log(`🚀 Completed in ${usageTime}s`)
 }
